clean up and simplify video recorder logic

+++ src/app/video-recorder.ts
@@ -1,11 +1,8 @@
 import { BrowserWindow, shell } from "electron"
 import type { ExecaChildProcess } from "execa"
-import { createWriteStream } from "fs"
-import { mkdir } from "fs/promises"
 import { observable } from "micro-observables"
-import { unlink } from "node:fs/promises"
-import { dirname } from "path"
-import { pipeline } from "stream/promises"
+import { mkdir, unlink } from "node:fs/promises"
+import { dirname } from "node:path"
 import { z } from "zod"
 import { name as appName } from "../../package.json"
 import { isDev } from "../common/constants"
@@ -15,14 +12,16 @@ import { getDistPath } from "../common/paths"
 import type { Rect } from "../common/Rect"
 import { rect } from "../common/Rect"
 import { vec } from "../common/Vec"
-import { importExeca } from "./execa"
 import { getVideoRecordingsPath } from "./paths"
 import { tryRegisterShortcut } from "./try-register-shortcut"
+
+// execa is an ES module, and can't be required normally
+const execaPromise = import("execa")
 
 export type VideoRecorderState =
   | { status: "ready" }
   | { status: "preparing" }
-  | { status: "recording"; process: ExecaChildProcess; outputPath: string }
+  | { status: "recording"; process: VideoRecorderProcess; outputPath: string }
 
 export class VideoRecorder {
   readonly state = observable<VideoRecorderState>({ status: "ready" })
@@ -59,35 +58,35 @@ export class VideoRecorder {
     if (this.state.get().status !== "ready") return
     this.state.set({ status: "preparing" })
 
-    const outputPath = VideoRecorder.getRecordingOutputPath()
+    let outputPath: string | undefined
 
     try {
-      const region = await VideoRecorder.getRegion()
+      outputPath = await VideoRecorder.ensureRecordingOutputPath()
+      const region = await VideoRecorder.selectRegion()
+      const process = new VideoRecorderProcess()
+
+      this.state.set({ status: "recording", process, outputPath })
       this.showWindows(region)
 
-      const [child] = await VideoRecorder.runRecordingProcess(region)
+      await process.start(region, outputPath)
 
-      this.state.set({ status: "recording", process: child, outputPath })
-
-      await mkdir(dirname(outputPath), { recursive: true })
-      await pipeline([child.stdout!, createWriteStream(outputPath)])
+      shell.showItemInFolder(outputPath)
     } catch (error) {
-      if (await isFile(outputPath)) {
+      if (outputPath && (await isFile(outputPath))) {
         await unlink(outputPath)
       }
       throw error
     } finally {
       this.state.set({ status: "ready" })
+      this.hideWindows()
     }
   }
 
   stopRecording() {
     const state = this.state.get()
-    if (state.status !== "recording") return
-    state.process.stdin!.write("q")
-    shell.showItemInFolder(state.outputPath)
-    this.hideWindows()
-    this.state.set({ status: "ready" })
+    if (state.status === "recording") {
+      state.process.stop()
+    }
   }
 
   private static async createRecordingFrameWindow() {
@@ -125,20 +124,8 @@ export class VideoRecorder {
     return win
   }
 
-  private static getRecordingOutputPath() {
-    const timestamp = [
-      new Date().getFullYear(),
-      new Date().getMonth() + 1,
-      new Date().getDate(),
-      new Date().getHours(),
-      new Date().getMinutes(),
-      new Date().getSeconds(),
-    ].join("-")
-    return getVideoRecordingsPath(`${appName}-${timestamp}.avi`)
-  }
-
-  private static async getRegion(): Promise<Rect> {
-    const execa = await importExeca()
+  private static async selectRegion(): Promise<Rect> {
+    const { execa } = await execaPromise
 
     const result = await execa("slop", [
       "--highlight",
@@ -154,33 +141,23 @@ export class VideoRecorder {
     })
 
     const region = regionSchema.parse(JSON.parse(result.stdout))
+
     return rect(vec(region.x, region.y), vec(region.width, region.height))
   }
 
-  private static async runRecordingProcess(
-    region: Rect,
-  ): Promise<[ExecaChildProcess<string>]> {
-    const args = [
-      // global options
-      `-f x11grab`,
-      `-loglevel warning`,
+  private static async ensureRecordingOutputPath() {
+    const timestamp = [
+      new Date().getFullYear(),
+      new Date().getMonth() + 1,
+      new Date().getDate(),
+      new Date().getHours(),
+      new Date().getMinutes(),
+      new Date().getSeconds(),
+    ].join("-")
 
-      // input / x11grab options
-      `-framerate 30`,
-      `-video_size ${region.width}x${region.height}`,
-      `-i ${process.env.DISPLAY || ":0"}.0+${region.left},${region.top}`,
-
-      // output options
-      `-qscale 0`,
-      `-vcodec huffyuv`,
-      `-f avi`,
-      `-`,
-    ]
-
-    const execa = await importExeca()
-    const child = execa("ffmpeg", args.map((it) => it.split(/\s+/)).flat())
-    child.catch(logErrorStack)
-    return [child]
+    const path = `${appName}-${timestamp}.mp4`
+    await mkdir(dirname(path), { recursive: true })
+    return getVideoRecordingsPath(path)
   }
 
   private showWindows(region: Rect) {
@@ -196,5 +173,44 @@ export class VideoRecorder {
 
   private hideWindows() {
     this.regionFrame.hide()
+  }
+}
+
+class VideoRecorderProcess {
+  private child?: ExecaChildProcess
+
+  async start(region: Rect, outputPath: string) {
+    const { execaCommand } = await execaPromise
+
+    this.child = execaCommand(
+      [
+        "ffmpeg",
+
+        // global options
+        `-f x11grab`,
+        `-hide_banner`,
+
+        // input options
+        `-framerate 30`,
+        `-video_size ${region.width}x${region.height}`,
+        `-i ${process.env.DISPLAY || ":0"}.0+${region.left},${region.top}`,
+
+        // output options
+        // I've tried using x265,
+        // but that outputs the x265 cli banner in the output file itself,
+        // which corrupts it for certain video players,
+        // so sticking with x264 for now
+        `-vcodec libx264`,
+        outputPath,
+      ].join(" "),
+      { stdout: "inherit" },
+    )
+
+    await this.child
+  }
+
+  stop() {
+    this.child?.stdin!.write("q")
+    this.child = undefined
   }
 }
